Refactor components and add page meta data
Refactor the "about" and "[slug]" page components. Removed the import of the "useStateContext" hook in the "about.js" file and removed the "toast" import in the "[slug].js" file. Added a new import for the "PageMeta" component in the "[slug].js" file. Modified the "Item" component by adding a new "PageMeta" component with title and description props. Added a new file "stand/[slug].js" with React hook and import. Made styling changes in the "status.sass" file by adding a "z-index" property to the ".status-heading" class.