feat(api): Add type parameter to getDataBySlug function
This commit adds a new `type` parameter to the `getDataBySlug` function in the `cosmic.js` file. Previously, the function only accepted the `slug` parameter. Now, it also accepts the `type` parameter, which defaults to `'products'`. This allows for more flexibility in fetching data by slug, as different types can be specified.

It also removes unnecessary code in the `Item` component in the `item/[slug].js` file. The `onAdd`, `cartItems`, and `cosmicUser` variables are no longer used in the component and have been removed.

Lastly, it updates the `stand/[slug].js` file by importing the `Layout` component, `getAllDataByType`, `getDataBySlug` functions from the `cosmic.js` file, and the `PageMeta` component from the `Meta` file. It also adds a `PageMeta` component and replaces the component's rendering logic to use the imported components.

These modifications improve the functionality, organization, and structure of the codebase.