feat: Add additional properties to getDataByCategory and getAllDataByType functions
This change modifies the `getDataByCategory` and `getAllDataByType` functions in the `cosmic.js` file. It adds additional properties (`type`, `created_by`, `modified_by`, and `created_at`) to the list of properties that are retrieved from the Cosmic JS API. This change increases the amount of data returned by these functions, providing more information about the objects retrieved.

Fix access condition in Item component

In the `item/[slug].js` file, a condition for displaying the delete button based on the user's ID and the item's `modified_by` property has been updated. The condition now checks for the user's email instead of the ID when comparing against the item's `metadata.email` property. This change ensures that the delete button is shown correctly when the user's email matches the email stored in the item's metadata.