Implement GitHub action for njsscan integration with Code Scanning feature
This commit adds a GitHub action workflow file `.github/workflows/njsscan.yml`. The workflow integrates `njsscan` with GitHub's Code Scanning feature. `njsscan` is a static security code scanner that identifies insecure code patterns in Node.js applications. The workflow is triggered on `push` and `pull_request` events to the `dev` branch, as well as on a weekly schedule. The workflow includes necessary permissions to fetch code, run security scanning, and upload SARIF results. The code is checked out using the `actions/checkout@v3` action, and the `ajinabraham/njsscan-action` is used to