Add .editorconfig file to project
Adds a new file `.editorconfig` to the project root directory. This file specifies formatting rules for all files in the project. The `.editorconfig` file sets the indentation style to space with a size of 2, specifies UTF-8 charset, and defines that trailing whitespace should be trimmed and a final newline should be inserted at the end of each file.