feat: Add keyboard accessibility to User dropdown menu
Add keyboard accessibility to the User dropdown menu in the Header/User component. This allows users to toggle the visibility of the dropdown menu by pressing the Enter key. Additionally, the menu items in the dropdown menu can now be selected using the keyboard by pressing the Enter key. This improves the overall accessibility and user experience of the component.

Refactor Headers component

Refactor the Headers component in order to simplify the logic and improve readability. The callback function handleOAuth now uses optional chaining and object property shorthand to check if the cosmicUser object does not have a property "id" and the user object does have a property "id". If the condition is true, the cosmicUser state is updated with the user object. This simplifies the code and makes it more concise.

Refactor HotBid component

Refactor the HotBid component to improve code readability and maintainability. The SlickArrow component now explicitly passes down necessary props to the button element. The button element is also now given a type attribute of "button" to indicate its purpose. This enhances the accessibility of the component.

The Hot component has been modified to use a unique identifier, such as item.id, instead of the index as the key prop when mapping over the info array. This ensures that each rendered Card component has a unique key, which is important for React's reconciliation algorithm to work correctly.