feat: Update .eslintrc.json
This commit updates the `.eslintrc.json` file. It adds three new values to the `"extends"` property - `"next"`, `"next/core-web-vitals"`, and `"airbnb"`. It also updates the `"ecmaVersion"` property to `2023`.

refactor: Update AppLink component

This commit updates the `AppLink` component. It adds `PropTypes` import and declares prop types for `href`, `className`, and `children` props. It also adds default values for these props.

refactor: Update Group component in Footer

This commit updates the `Group` component in the `Footer` folder. It updates the error handling by changing `Error(error)` to `Error(error.status)`.

refactor: Update Headers component

This commit updates the `Headers` component. It adds an `if` statement with a conditional to check if `cosmicUser` object has a property `"id"`. If it doesn't, and `user` object has a property `"id"`, it sets the `CosmicUser` state with the value of `user`.

refactor: Update HotBid component

This commit updates the `HotBid` component. It adds `PropTypes` import and declares prop types for `currentSlide`, `slideCount`, and `children` props. It also adds default values for these props.