Add CodeQL workflow file
This commit adds a new workflow file for CodeQL analysis. The workflow is triggered on push and pull_request events on the "dev" branch, as well as on a schedule. The workflow includes steps to analyze the code using CodeQL and perform security evaluations. It also initializes the CodeQL tools and uses autobuild to build any compiled languages. The analysis is categorized based on the language detected in the repository.