Refactor image sizing logic and fix width and height values
The commit modifies the code in the `ImageApp` component to refactor the logic for determining the width and height of the image.

Previously, if the width or height value was set to '100%', it would be replaced with a hardcoded value of '1000'. However, this behavior has been changed. Now, if the width or height is not an integer, it will be fixed with the provided numeric value defined in the `newSize` object.

Additionally, the code has been updated to set the `width` and `height` CSS properties of the image's style to the original values of `size.width` and `size.height` respectively, instead of using hardcoded '100%' values.

These modifications improve the flexibility and maintainability of the image sizing logic in the `ImageApp` component.