feat: Update deleteHandler to include a success message
Updated the deleteHandler function in delete.js to include a success message when the gift is successfully deleted. Previously, only the response data was returned. Now, a JSON object containing the data and the message "Le cadeau a bien été supprimé" is sent as the response.

In addition, added error handling logic to include the entire error object in the response when an error occurs.

In [slug].js, made a modification to the Item component. When the handleOAuth function is called and the cosmicUser.id is not present, added an "await" keyword to ensure that the handleOAuth function completes before proceeding.

Also, updated the fetch request to the '/api/delete' endpoint. Removed the JSON.stringify() call when passing the idProduct in the request body and removed the 'Content-Type' header from the request.

Finally, added a conditional check for the presence of a deleteItem.message and the value of deleteItem.messageOK being "Le cadeau a bien été supprimé". If these conditions are met, a success toast message is displayed and a redirect to '/search' is scheduled after 3 seconds.