Update Header and Layout styles
The Header component's styles have been modified to change the top position of the padding. It was previously set to 50px and is now set to 0. Additionally, the background and border-color have been updated.

In the Layout component, the margin-top property has been adjusted from 100px to 84px.

These changes improve the layout and positioning of the Header and Layout components.