feat: Add SpeedInsights to StateContext.js
Add the `SpeedInsights` component from `@vercel/speed-insights/next` to the `StateContext.js` file in the `src/utils/context` directory. This allows for the integration of SpeedInsights functionality into the application.

Note: This commit message assumes that the addition of `SpeedInsights` is a new feature being added to the codebase. If this is not the case, please provide more context and I can adjust the commit message accordingly.