DeleteHandler: Handle response for delete operation
This commit modifies the deleteHandler function in `src/pages/api/delete.js`. It updates the implementation of the `deleteHandler` by making changes to the response returned based on the input received. Previously, the function would delete the `req.body` object and return a successful deletion message. Now, it checks if `req.body.creator` is present. If yes, it returns a success message with additional data (`data.id` and `messageOK`), otherwise, it returns an error message with a status code of 401 (`Access Denied, you are not the creator`). This change ensures a more specific response based on the input received during the delete operation.

Item: Display delete success toast message

This commit modifies the `Item` component in `src/pages/item/[slug].js`. It updates the logic to display a success toast message when the `deleteItem` operation is successful. Previously, it would check if `deleteItem.data.message` was present and if `deleteItem.messageOK` was equal to 'Le cadeau a bien été supprimé'. Now, it checks if `deleteItem.messageOK` is present and if it's equal to 'Le cadeau a bien été supprimé'. If the conditions are met, a success message is displayed using `toast.success()`, and after a delay of 3000 milliseconds, it navigates to the '/search' page. This change provides improved feedback to the user when a deletion is successful.