Disable `linebreak-style` rule in `.eslintrc.json`
This change disables the `linebreak-style` rule in the `.eslintrc.json` file. This rule is used by ESLint to enforce consistent linebreaks in the codebase. By setting it to `0`, we are indicating that this rule should be turned off. This modification allows for more flexibility in the linebreaks used in the code.